removes authenticate() call from NewClient
i just cannot get past the icky feeling i have about making a net call
whenever creating a new object. we can add the token check and auth to
the CreateAttack func.
